Creating the "nft-gallery" repository. Version 1.0.0.
js/jsx - Work on the "Header" component is finished. The "Framer Motion" library was added to the component. The basic variables for the animations were prescribed.

css/scss - The main files to work with styles (Import.scss and Variables.scss) were added. Files App.css and Header.css, received the basic properties of styles and flex elements.

img - The main images were added to work with the component "Header".

json - The libraries "framer-motion", "axios" and "react-scroll" were added to work with the project.