Version 1.0.2. Works on footer
js/jsx - Created component Footer. Was made the initial blanks to work with the component.

css/scss - Footer.scss/css files were created. In the App.scss/css file, were added changes for the overall width of the site.

img - The main applications, to work with the component, were added