Version 1.1.1. Work on Collection.jsx.
js/jsx - Collection.jsx component was added to the main component App.jsx, Collection.jsx/css files were imported. Slightly reworked the file structure and changed the documentation. In the component Collection.jsx, were imported images to work with the project. In tags and classes, content was added. Stubs were added for links.

css/scss - in the file Collection.scss/css, were added to the basic properties of styles, grid and flex elements. Styles were added for common classes. Hover animations were added to the content.

img - An image was added to work with the component.