Version 2.0.1.  Working on the Seller.jsx component.
js/jsx - Work on the Seller.jsx component was started. The first part of the component was made, work on the second part began. The css-modules were added and used.

css/scss - Basic properties of styles, flex and grid elements, animations when hovering with content were added.

img - images were added to work with the component.