Version 2.1.3. The Users.jsx component is now complete.
js/jsx - In the Users.jsx component, css-modules for the main tags were added. The content was added to the tags. In the process, the code was slightly reworked.

css/scss - Basic CSS style properties from the layout were added, grid and flex elements were added. With the help of pseudoclasses, animations were added when hovering over the content.

img - Было добавлено изображение, для работы с компонентом.