Version - Version 2.1.4. Work has begun on the Blog.jsx component.
js/jsx - Work on the component was started, it was added to App.jsx. Basic tags were added to the component.

css/scss - In the file scss/css, were added to the main classes and subclasses to work with the component.

img - Images were added to work with the component.